Make the state-average solver state-specific but with common truncation of all states
Convert the state-average solver (sa_algorithm = -1) to state-specific, but retain
a common SVD truncation for all states.
This is closer to the original non-MPS SA-DMRG formulation by Chan et al., where states
are optimized individually, and then an average density matrix is constructed and
used for truncation of all the states.
The main purpose of such solver is to retain the common truncation for all states and to
ensure the local basis for all states is the same.